weekly-report-reviewer/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
line-bot-sdk>=3.0.0
